mail-followup-gradio/requirements.txt
gradio
pymupdf
langchain
openai
python-dotenv
gradio